src/sql_bigbrother/core/backup/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
mysql-connector-python==8.4.0
python-dotenv==1.0.1
langchain==0.1.20
langchain-community==0.0.38
crewai==0.30.11